requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.51.0
ecdsa==0.18.0
idna==3.4
oauthlib==3.2.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
